feat: changed Stelle in Feedback, from int to float

+++ backend/database/migrations/2024_08_17_172651_feedback.php
@@ -16,7 +16,7 @@ return new class extends Migration
                 `Matricola` int NOT NULL,
                 `idUniversita` int NOT NULL,
                 `idTirocinio` int NOT NULL,
-                `Stelle` int NOT NULL CHECK(Stelle BETWEEN 1 AND 5),
+                `Stelle` float NOT NULL CHECK(Stelle BETWEEN 1 AND 5),
                 `Descrizione` varchar(255) NOT NULL,
                 PRIMARY KEY (`Matricola`, `idUniversita`, `idTirocinio`),
                 FOREIGN KEY (`Matricola`, `idUniversita`) REFERENCES `AccountStudente` (`Matricola`, `idUniversita`),
